+++ lib/govuk_personalisation/test_helpers/requests.rb
@@ -1,8 +1,8 @@
 module GovukPersonalisation
   module TestHelpers
     module Requests
-      def mock_logged_in_session(value = "placeholder")
-        request.headers["GOVUK-Account-Session"] = value
+      def mock_logged_in_session(value = "placeholder", flash = nil)
+        request.headers["GOVUK-Account-Session"] = GovukPersonalisation::Flash.encode_session(value, flash)
       end
     end
   end